requirements.txt
beautifulsoup4==4.12.2
lxml_html_clean
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
click==8.1.7
dateparser==1.2.0
distro==1.8.0

djangorestframework
django-cors-headers
Flask==2.2.5
google==3.0.0
googlesearch-python==1.2.3
gunicorn==21.2.0
htmldate==1.6.0
httpcore==1.0.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
libretranslatepy==2.1.1
lxml>=4.9.2,<4.10.0
MarkupSafe==2.1.2
msgpack==1.0.5
nltk
numpy
packaging==23.2
pydantic
pydantic_core
python-dateutil
python-dotenv
pyttsx3==2.90
redis==5.0.1
regex
requests
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SpeechRecognition==3.9.0
sqlparse==0.4.4
threadpoolctl==3.1.0
tqdm==4.66.1
translate==3.6.1
typing-extensions
tzlocal==5.1
urllib3==2.0.7
Werkzeug==2.2.3
whitenoise
scikit-learn
zipp==3.15.0
dj-database-url
# psycopg2
psycopg2-binary
tavily-python
retry
openai==1.55.3
httpx==0.27.2
langchain
langchain_openai
langchain_community
django-autoslug
langchain_groq
langchain_tavily